pyproject.toml
#:schema None
##:schema https://json.schemastore.org/pyproject.json

[project]
name = "archivepodcast"
version = "1.5.0"
description = "Flask webapp that will archive a podcast from a RSS feed. It will download the episodes and re-host them."
authors = [{ name = "Kieran Gee", email = "kieran.lost.the.game@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "flask[async]>=3.1,<4",
    "aiobotocore>=2.25.2",
    "aiohttp>=3.13.2",
    "jinja2>=3.1,<4",
    # dnf install libxml2-devel libxslt-devel
    # apt install libxml2-dev libxslt-dev
    "lxml>=5,<5.4", # Right now its broken for linux on a greater version
    # "lxml>=6.0.2", # TESTING
    "markdown>=3.10",
    "psutil>=7.1.3",
    "pydantic>=2.12",
    "pydantic-settings>=2.12",
    "python-magic>=0.4",
    "rich>=14.2.0",
    "typed-ffmpeg>=3.7",
    "waitress>=3.0",
]

[project.optional-dependencies]
type = [
    "mypy",
    "ty",
    "types-lxml",
    "types-psutil",
    "types-Markdown",
    "botocore-stubs",
    "types-aiobotocore",
    "types-aiobotocore-s3",
]

lint = ["ruff"]

profile = ["aiomonitor"]

test = [
    "pytest>=9.0",
    "pytest-cov",
    "pytest-mock",
    "pytest-random-order",
    "detect-test-pollution",
    "types-aiobotocore-s3", # Mock uses typed dict
    "pytest-asyncio",
    "tomlkit>=0.13",
    "pytest-socket>=0.7.0",
]

docs = ["sphinx", "myst-parser", "sphinx-rtd-theme"]


[tool.ruff]
cache-dir = "~/.cache/ruff"
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"] # KG Include absolutely all of Ruff's rules.
ignore = [
    # Modules
    "DJ", # KG Django not required

    # Specific rules
    "TD002",  # KG Verbose TODO, only I work on this project
    "TD003",  # KG Verbose TODO, only I work on this project
    "S311",   # KG I'll assume no real crypto will be done.
    "ISC001", # KG conflict in the formatter
    "COM812", # KG conflict in the formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    # Modules
    "ARG", # KG Unused function args -> fixtures nevertheless are functionally relevant
    "FBT", # KG Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    "D",   # KG Don't care about docstrings in tests
    "ANN", # KG Don't care about annotations in tests

    # Specific rules
    "S101",    # KG asserts allowed in tests.
    "SLF001",  # KG Accessing private members is needed for testing.
    "PT004",   # KG monkeypatch doesn't return.
    "S105",    # KG Hardcoded fake passwords are fine in tests.
    "PLR0913", # KG Tests can have as many arguments as they want.
    "BLE001",  # KG Tests can use BaseException.
]
"tests/fixtures/aws.py" = [
    "N803", # KG AWS Just breaks the rules here
]

"docs/*.py" = [
    # Modules
    "INP",
    "D",
    "A",
]


[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 40

[tool.pytest.ini_options]
addopts = [
    "--cov-context=test",
    "--cov=archivepodcast",
    "--cov-report=term",
    "--cov-report=html",
    "--disable-socket",
    "--allow-hosts=127.0.0.1,127.0.1.1",
    "--asyncio-mode=auto",
]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:boto.*",
    "ignore::pytest.PytestUnhandledThreadExceptionWarning", # God knows
]


[tool.coverage.html]
show_contexts = true


[tool.mypy]
files = ["archivepodcast", "tests"]
namespace_packages = false # God knows why this makes mypy take forever
show_error_codes = true
strict = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
    "type-arg",
    "unused-awaitable",
    "possibly-undefined",
    "redundant-self",
]
disable_error_code = ["misc"]
warn_unreachable = true
warn_redundant_casts = true


[[tool.ty.overrides]]
# Blame the flask client object
include = ["tests/**"]
[tool.ty.overrides.rules]
"unresolved-attribute" = "ignore"

[[tool.ty.overrides]]
# Not sure what ty is upset about here
include = [
    "archivepodcast/blueprints/rss.py",
    "archivepodcast/archiver/podcast_archiver.py",
]
[tool.ty.overrides.rules]
"invalid-argument-type" = "ignore"
